+++ tests/testthat/test-parallel.r
@@ -1,15 +1,15 @@
 context("Parallel")
 
-if (require("doMC", quietly = TRUE)) {
-  registerDoMC(2)
+if (require("doParallel", quietly = TRUE)) {
+  registerDoParallel(cores = 2)
 
-  if (identical(Sys.getenv("NOT_CRAN"), "true")) {
-    test_that("l_ply respects .parallel", {
-      expect_that(
-        l_ply(c(0.1, 0.1), Sys.sleep, .parallel = TRUE),
-        takes_less_than(0.15))
-    })
-  }
+  test_that("l_ply respects .parallel", {
+    skip_on_cran()
+
+    expect_that(
+      l_ply(c(0.1, 0.1), Sys.sleep, .parallel = TRUE),
+      takes_less_than(0.15))
+  })
 
   test_that("l_ply + .parallel complains about invalid arguments", {
     expect_message(
@@ -27,5 +27,5 @@ if (require("doMC", quietly = TRUE)) {
     expect_equal(x, NULL)
   })
 
-  registerDoMC(1)
+  registerDoParallel(1)
 }