+++ R/parallel.r
@@ -1,16 +1,13 @@
 setup_parallel <- function() {
   if (!requireNamespace("foreach", quietly = TRUE)) {
+    # EXCLUDE COVERAGE START
     stop("foreach package required for parallel plyr operation",
       call. = FALSE)
+    # EXCLUDE COVERAGE END
   }
   if (foreach::getDoParWorkers() == 1) {
+    # EXCLUDE COVERAGE START
     warning("No parallel backend registered", call. = TRUE)
+    # EXCLUDE COVERAGE END
   }
-}
-
-parallel_fe <- function(n, options) {
-  i <- seq_len(n)
-  fe_call <- as.call(c(list(as.name("foreach"), i = i), options))
-
-  eval(fe_call)
 }