+++ R/plyr-deprecated.r
@@ -1,0 +1,14 @@
+#' Deprecated Functions in Package plyr
+#'
+#' These functions are provided for compatibility with older versions of
+#' \code{plyr} only, and may be defunct as soon as the next release.
+#'
+#' \itemize{
+#'   \item \code{\link{liply}}
+#'   \item \code{\link{isplit2}}
+#' }
+#'
+#' @name plyr-deprecated
+# EXCLUDE COVERAGE START
+NULL
+# EXCLUDE COVERAGE END