+++ NEWS.md
@@ -8,6 +8,9 @@
   `.id = NULL` to work as described in the help. (Thanks to Doug Mitarotonda, #207, 
   and Marek, #224 and #225)
 
+* Warn on duplicate names that cause certain array functions to fail. (Thanks
+  to Kirill Müller, #211)
+
 # Version 1.8.1
 
 * New parameter `.id` to `ldply()` and `rdply()` that specifies the name of